updating interface + files

+++ interface/interf.py
@@ -3,13 +3,14 @@ from tkinter import *
 from tkinter import ttk
 import os
 import time
+from tkinter import messagebox
 
 
 def client_tube(second_frame, gui2,frame, btn):
    gui2.geometry("635x490")
    btn.pack(ipady=0, ipadx=0,expand=False)
    frame.pack(expand=True, fill=BOTH, padx=20, pady=20)
-   Commande="\" ../tube_nomme/client >> ClientsTubeNomme.txt; exec bash\""
+   Commande="\" ../tube_nomme/executables/client >> ClientsTubeNomme.txt; exec bash\""
    subprocess.Popen(f"gnome-terminal --tab -- bash -c {Commande}",shell=True)
    time.sleep(3)
    f = open("ClientsTubeNomme.txt")
@@ -25,52 +26,59 @@ def client_tube(second_frame, gui2,frame, btn):
 
 def client_socket(gui2,e):
                 
-   nb=int(e.get())   
-   #create new window
-   gui3 = Toplevel(gui2)
-   gui3.geometry("650x450")  
-   # Add image file
-   bg = PhotoImage(file = "img.png")
-   label1 = Label( gui3, image = bg)
-   label1.place(x = 0, y = 0)
+   nb=int(e.get())
+   if nb>=1 and nb<=15:
+      #create new window
+      gui3 = Toplevel(gui2)
+      gui3.geometry("650x490")  
+      # Add image file
+      bg = PhotoImage(file = "img.png")
+      label1 = Label( gui3, image = bg)
+      label1.place(x = 0, y = 0)
 
-   frame = Frame(gui3,bg="#4a6f7c")
-   frame.pack(expand=True, fill=BOTH, padx=20, pady=20)
-   # Create A Canvas
-   my_canvas = Canvas(frame,bg="#4a6f7c")
-   my_canvas.pack(side=LEFT, fill=BOTH, expand=1)
+      frame = Frame(gui3,bg="#99bbc7")
+      frame.pack(expand=True, fill=BOTH, padx=20, pady=20)
+      # Create A Canvas
+      my_canvas = Canvas(frame,bg="#99bbc7")
+      my_canvas.pack(side=LEFT, fill=BOTH, expand=1)
 
-   # Add A Scrollbar To The Canvas
-   my_scrollbar = ttk.Scrollbar(frame, orient=VERTICAL, command=my_canvas.yview)
-   my_scrollbar.pack(side=RIGHT, fill=Y)
+      # Add A Scrollbar To The Canvas
+      my_scrollbar = ttk.Scrollbar(frame, orient=VERTICAL, command=my_canvas.yview)
+      my_scrollbar.pack(side=RIGHT, fill=Y)
 
-   # Configure The Canvas
-   my_canvas.configure(yscrollcommand=my_scrollbar.set)
-   my_canvas.bind('<Configure>', lambda e: my_canvas.configure(scrollregion = my_canvas.bbox("all")))
+      # Configure The Canvas
+      my_canvas.configure(yscrollcommand=my_scrollbar.set)
+      my_canvas.bind('<Configure>', lambda e: my_canvas.configure(scrollregion = my_canvas.bbox("all")))
 
-   # Create ANOTHER Frame INSIDE the Canvas
-   second_frame = Frame(my_canvas,bg="#4a6f7c")  
-   my_canvas.create_window((0,0), window=second_frame, anchor="nw")  
+      # Create ANOTHER Frame INSIDE the Canvas
+      second_frame = Frame(my_canvas,bg="#99bbc7")  
+      my_canvas.create_window((0,0), window=second_frame, anchor="nw")  
 
-   
-   for i in range(nb):
-      Commande="\" ../Sockets/client >> ClientsSockets.txt; exec bash\""
+      Commande="\" ../Sockets/executables/client >> ClientsSockets.txt"      
+      for i in range(nb-1):
+         Commande = Commande + " & ../Sockets/executables/client >> ClientsSockets.txt"
+      Commande = Commande + "; exec bash\""
+         
       subprocess.Popen(f"gnome-terminal --tab -- bash -c {Commande}",shell=True)
       time.sleep(3)
       f = open("ClientsSockets.txt")
-      Label(second_frame, text="Client "+str(i+1),bg="#4a6f7c",fg="#760523",font= ('Helvetica 15 bold underline')).pack()
-
+      #Label(second_frame, text="Client "+str(i+1),bg="#99bbc7",fg="#760523",font= ('Helvetica 15 bold underline')).pack()
       lines=f.readlines()
       for line in lines:
          temp_text = line
-         Label(second_frame, text=temp_text,bg="#4a6f7c",font= ('italic 11 bold')).pack()
+         if temp_text[0] == "*":
+            Label(second_frame, text=temp_text, fg="#5c7d89", bg="#99bbc7",font= ('italic 11') ).pack()
+         else:     
+            Label(second_frame, text=temp_text, fg="Black", bg="#99bbc7",font= ('italic 11 bold') ).pack()
       f.close()
       os.remove("ClientsSockets.txt")
-   gui3.mainloop()    
+      gui3.mainloop() 
+   else:
+      messagebox.showerror("Error","Le nombre de clients doit être entre 1 et 15")   
 
 def sel():
    if v.get()=="Tube":
-      Commande1="\" ../tube_nomme/server; exec bash\""
+      Commande1="\" ../tube_nomme/executables/server; exec bash\""
       subprocess.Popen(f"gnome-terminal --tab -- bash -c {Commande1}",shell=True)
       gui2 = Toplevel(gui)
       gui2.geometry("500x400")
@@ -109,12 +117,12 @@ def sel():
       label1 = Label( gui2, image = bg)
       label1.place(x = 0, y = 0)
    
-      l = Label(gui2, text = "Veuillez entrer le nombre de clients que vous désirez l'exécuter",bg="#4a6f7c",font= ('Helvetica 12 bold')).place(relx=.5, rely=.3,anchor= CENTER)
+      l = Label(gui2, text = "Veuillez entrer le nombre de clients que vous désirez l'exécuter",bg="#4a6f7c",font= ('Helvetica 13 bold')).place(relx=.5, rely=.3,anchor= CENTER)
       e = Entry(gui2, bd = 5)
       e.pack(pady= 155)
-      Commande1="\" ../Sockets/server; exec bash\""
+      Commande1="\" ../Sockets/executables/server; exec bash\""
       subprocess.Popen(f"gnome-terminal --tab -- bash -c {Commande1}",shell=True)
-      btn = Button(gui2, text ="Lancer",fg="#00222d",font= ('Helvetica 12 bold'),command=lambda:client_socket(gui2,e)).place(anchor=CENTER, relx=.5, rely=.5)
+      btn = Button(gui2, text ="Lancer",bg="#99bbc7",font= ('Helvetica 12 bold'),command=lambda:client_socket(gui2,e)).place(anchor=CENTER, relx=.5, rely=.5)
       gui2.mainloop()
           
 gui = Tk()
